Always render Modal, use open state
